third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/local-url-inherit-controller.https-expected.txt

This is a testharness.js-based test.
FAIL Same-origin blob URL iframe should inherit service worker controller. assert_equals: blob URL iframe should inherit controller expected (string) "https://web-platform.test:8444/service-workers/service-worker/resources/local-url-inherit-controller-worker.js" but got (object) null
FAIL Same-origin blob URL iframe should intercept fetch(). assert_equals: blob URL iframe should intercept fetch expected "intercepted" but got "Hello world\n"
FAIL Same-origin blob URL worker should inherit service worker controller. promise_test: Unhandled rejection with value: "Uncaught TypeError: Cannot read property 'controller' of undefined"
PASS Same-origin blob URL worker should intercept fetch().
FAIL Data URL iframe should not inherit service worker controller. assert_equals: data URL iframe should not inherit controller expected (object) null but got (string) "Failed to read the 'serviceWorker' property from 'Navigator': Access to service workers is denied in this document origin."
PASS Data URL iframe should not intercept fetch().
FAIL Data URL worker should not inherit service worker controller. promise_test: Unhandled rejection with value: "Uncaught TypeError: Cannot read property 'controller' of undefined"
FAIL Data URL worker should not intercept fetch(). assert_equals: data URL worker should not intercept fetch expected "" but got "intercepted"
Harness: the test ran to completion.
